requirements.txt.txt
python-telegram-bot==20.7
playwright